[#1027] [V71002] Merge GT.M V7.1-002 into YottaDB mainline (with conflicts)
* This commit was created using the following commands.

  ```
  git cherry-pick tags/V7.1-002
  git commit -an -S -m '[#1027] [V71002] Merge GT.M V7.1-002 into YottaDB mainline (with conflicts)'
  ```

* All merge conflicts are left as is in this commit and will be resolved manually in a future commit.

* No copyright changes done in this commit (hence the `-n` use in `git commit` above) since this
  was an automatically generated commit.

+++ sr_port/mvalconv.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- * Copyright (c) 2001-2022 Fidelity National Information	*
+ * Copyright (c) 2001-2023 Fidelity National Information	*
  * Services, Inc. and/or its subsidiaries. All rights reserved.	*
  *								*
  * Copyright (c) 2019-2024 YottaDB LLC and/or its subsidiaries. *
@@ -315,7 +315,7 @@ int4 mval2i(mval *v)
 		{
 			j = mval2double(v);
 			if (MINNEGINT4 > j)
-				i = MINNEGINT4;
+				i = (int4) MINNEGINT4;
 			else if (MAXPOSINT4 < j)
 				i = MAXPOSINT4;
 			else	/* if ((MINNEGINT4 <= j) && (MAXPOSINT4 >= j)) */
